Project structure (#32)
* Deleted two empty directories

* Renamed Swen2Project project to TourPlanner.Presentation and adjusted namespaces

* Added Business Project

* Added DataAccess project

* Added Models Project for universal stuff

* Removed Models from Presentation Layer
