+++ Swen2Project.sln
@@ -1,6 +1,12 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Swen2Project", "Swen2Project\Swen2Project.csproj", "{A5FE3F8A-3FFF-489F-A7C4-53CA3A063538}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TourPlanner.Presentation", "TourPlanner.Presentation\TourPlanner.Presentation.csproj", "{A5FE3F8A-3FFF-489F-A7C4-53CA3A063538}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TourPlanner.Business", "TourPlanner.Business\TourPlanner.Business.csproj", "{D6715F52-421E-4A3B-9078-ADB33EC85763}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TourPlanner.DataAccess", "TourPlanner.DataAccess\TourPlanner.DataAccess.csproj", "{10C93EE4-57A9-4200-A411-9D9979275E34}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TourPlanner.Models", "TourPlanner.Models\TourPlanner.Models.csproj", "{3A6E10EC-110E-4592-8D12-599645405500}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -12,5 +18,17 @@ Global
 		{A5FE3F8A-3FFF-489F-A7C4-53CA3A063538}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A5FE3F8A-3FFF-489F-A7C4-53CA3A063538}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A5FE3F8A-3FFF-489F-A7C4-53CA3A063538}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D6715F52-421E-4A3B-9078-ADB33EC85763}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D6715F52-421E-4A3B-9078-ADB33EC85763}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D6715F52-421E-4A3B-9078-ADB33EC85763}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D6715F52-421E-4A3B-9078-ADB33EC85763}.Release|Any CPU.Build.0 = Release|Any CPU
+		{10C93EE4-57A9-4200-A411-9D9979275E34}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{10C93EE4-57A9-4200-A411-9D9979275E34}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{10C93EE4-57A9-4200-A411-9D9979275E34}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{10C93EE4-57A9-4200-A411-9D9979275E34}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3A6E10EC-110E-4592-8D12-599645405500}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3A6E10EC-110E-4592-8D12-599645405500}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3A6E10EC-110E-4592-8D12-599645405500}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3A6E10EC-110E-4592-8D12-599645405500}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
