--- Swen2Project/TourPlanner.Models/Tour.cs
@@ -1,6 +1,0 @@
-﻿namespace Swen2Project.TourPlanner.Models;
-
-public class Tour
-{
-    
-}
