--- Swen2Project/TourPlanner.Presentation/App.xaml.cs
@@ -1,10 +1,0 @@
-﻿using System.Windows;
-
-namespace Swen2Project.TourPlanner.Presentation;
-
-/// <summary>
-/// Interaction logic for App.xaml
-/// </summary>
-public partial class App : Application
-{
-}
