--- Swen2Project/TourPlanner.Presentation/Controls/SearchBarControl.xaml.cs
@@ -1,11 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace Swen2Project.TourPlanner.Presentation.Controls;
-
-public partial class SearchBarControl : UserControl
-{
-    public SearchBarControl()
-    {
-        InitializeComponent();
-    }
-}
