--- Swen2Project/TourPlanner.Presentation/Controls/TourListControl.xaml.cs
@@ -1,11 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace Swen2Project.TourPlanner.Presentation.Controls;
-
-public partial class TourListControl : UserControl
-{
-    public TourListControl()
-    {
-        InitializeComponent();
-    }
-}
