--- Swen2Project/TourPlanner.Presentation/Controls/TourListControls/TourDescriptionControl.xaml
@@ -1,17 +1,0 @@
-﻿<UserControl x:Class="Swen2Project.TourPlanner.Presentation.Controls.TourListControls.TourDescriptionControl"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             mc:Ignorable="d"
-             d:DesignHeight="300" d:DesignWidth="300">
-    <Border Background="#F9F9F9" Style="{StaticResource CardStyle}" MinHeight="100">
-        <TextBlock TextWrapping="Wrap">
-            This scenic bicycle tour from Vienna to Salzburg follows along the Danube 
-            river for much of the route, offering beautiful views of the Austrian 
-            countryside. The path is mostly flat with some gentle hills as you 
-            approach Salzburg. Suitable for intermediate cyclists, with plenty of 
-            villages and towns along the way for rest stops.
-        </TextBlock>
-    </Border>
-</UserControl>
