--- Swen2Project/TourPlanner.Presentation/Controls/TourListControls/TourDescriptionControl.xaml.cs
@@ -1,11 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace Swen2Project.TourPlanner.Presentation.Controls.TourListControls;
-
-public partial class TourDescriptionControl : UserControl
-{
-    public TourDescriptionControl()
-    {
-        InitializeComponent();
-    }
-}
