--- Swen2Project/TourPlanner.Presentation/Controls/TourListControls/TourInformationControl.xaml.cs
@@ -1,11 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace Swen2Project.TourPlanner.Presentation.Controls.TourListControls;
-
-public partial class TourInformationControl : UserControl
-{
-    public TourInformationControl()
-    {
-        InitializeComponent();
-    }
-}
