--- Swen2Project/TourPlanner.Presentation/Controls/TourListControls/TourStatisticControl.xaml.cs
@@ -1,11 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace Swen2Project.TourPlanner.Presentation.Controls.TourListControls;
-
-public partial class TourStatisticControl : UserControl
-{
-    public TourStatisticControl()
-    {
-        InitializeComponent();
-    }
-}
