--- Swen2Project/TourPlanner.Presentation/Views/AddingTourWindow.xaml.cs
@@ -1,11 +1,0 @@
-﻿using System.Windows;
-
-namespace Swen2Project.TourPlanner.Presentation.Views;
-
-public partial class AddingTourWindow : Window
-{
-    public AddingTourWindow()
-    {
-        InitializeComponent();
-    }
-}
