--- Swen2Project/TourPlanner.Presentation/Views/MainWindow.xaml
@@ -1,66 +1,0 @@
-﻿<Window x:Class="Swen2Project.TourPlanner.Presentation.Views.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:viewModels="clr-namespace:Swen2Project.TourPlanner.ViewModels"
-        xmlns:controls="clr-namespace:Swen2Project.TourPlanner.Presentation.Controls"
-        d:DataContext="{d:DesignInstance Type=viewModels:MainWindowViewModel}"
-        mc:Ignorable="d"
-        Title="Tour Planner" Height="700" Width="1200"
-        MinHeight="600" MinWidth="900"
-        Background="#F5F5F5">
-    <Window.Resources>
-        
-    </Window.Resources>
-    
-    <DockPanel>
-        <!-- Top Menu Bar for common commands -->
-        <Menu DockPanel.Dock="Top" Background="#2E7D32" Foreground="White" FontSize="13">
-            <MenuItem Header="_File" Foreground="White">
-                <MenuItem Header="_New Tour" />
-                <MenuItem Header="_Import Tours" />
-                <MenuItem Header="_Export Tours" />
-                <Separator/>
-                <MenuItem Header="_Exit" />
-            </MenuItem>
-            <MenuItem Header="_Edit" Foreground="White">
-                <MenuItem Header="_Edit Tour" />
-                <MenuItem Header="_Delete Tour" />
-            </MenuItem>
-            <MenuItem Header="_Reports" Foreground="White">
-                <MenuItem Header="_Tour Report" />
-                <MenuItem Header="_Summary Report" />
-            </MenuItem>
-            <MenuItem Header="_Help" Foreground="White">
-                <MenuItem Header="_About" />
-            </MenuItem>
-        </Menu>
-        
-        <!-- Status Bar at the bottom -->
-        <StatusBar DockPanel.Dock="Bottom" Background="#EEEEEE" Height="24" MinHeight="24">
-            <TextBlock Text="Ready" Margin="5,0" />
-        </StatusBar>
-        
-        <!-- Main Content Area -->
-        <Grid Margin="10">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="350" MinWidth="250"/>
-                <ColumnDefinition Width="*" MinWidth="400"/>
-            </Grid.ColumnDefinitions>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="Auto" MinHeight="60"/>
-                <RowDefinition Height="*" MinHeight="400"/>
-            </Grid.RowDefinitions>
-            
-            <!-- Search Bar -->
-            <controls:SearchBarControl Grid.Row="0" Grid.Column="0" Grid.ColumnSpan="2"></controls:SearchBarControl>
-            
-            <!-- Left Panel: Tours List -->
-            <controls:TourListControl Grid.Row="1" Grid.Column="0" />
-            
-            <!-- Right Panel: Tour Details -->
-            <controls:TourDetailsControl Grid.Row="1" Grid.Column="1"></controls:TourDetailsControl>
-        </Grid>
-    </DockPanel>
-</Window>
