--- Swen2Project/TourPlanner.Presentation/Views/MainWindow.xaml.cs
@@ -1,17 +1,0 @@
-﻿using System.Windows;
-using Swen2Project.TourPlanner.ViewModels;
-
-namespace Swen2Project.TourPlanner.Presentation.Views;
-
-/// <summary>
-/// Interaction logic for MainWindow.xaml
-/// </summary>
-public partial class MainWindow : Window
-{
-    public MainWindow()
-    {
-        InitializeComponent();
-        var mainWindowViewModel = new MainWindowViewModel();
-        DataContext = mainWindowViewModel;
-    }
-}
