--- Swen2Project/TourPlanner.Presentation/Views/RemovingTourWindow.xaml.cs
@@ -1,11 +1,0 @@
-﻿using System.Windows;
-
-namespace Swen2Project.TourPlanner.Presentation.Views;
-
-public partial class RemovingTourWindow : Window
-{
-    public RemovingTourWindow()
-    {
-        InitializeComponent();
-    }
-}
