+++ TourPlanner.Business/Class1.cs
@@ -1,0 +1,5 @@
+﻿namespace TourPlanner.Business;
+
+public class Class1
+{
+}
