+++ TourPlanner.DataAccess/Class1.cs
@@ -1,0 +1,5 @@
+﻿namespace TourPlanner.DataAccess;
+
+public class Class1
+{
+}
