+++ TourPlanner.Models/Class1.cs
@@ -1,0 +1,5 @@
+﻿namespace TourPlanner.Models;
+
+public class Class1
+{
+}
