+++ TourPlanner.Presentation/AssemblyInfo.cs
