+++ TourPlanner.Presentation/Presentation/App.xaml.cs
@@ -1,0 +1,10 @@
+﻿using System.Windows;
+
+namespace TourPlanner.Presentation.Presentation;
+
+/// <summary>
+/// Interaction logic for App.xaml
+/// </summary>
+public partial class App : Application
+{
+}
