+++ TourPlanner.Presentation/Presentation/Controls/SearchBarControl.xaml
@@ -1,0 +1,22 @@
+﻿<UserControl x:Class="TourPlanner.Presentation.Presentation.Controls.SearchBarControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             mc:Ignorable="d"
+             d:DesignHeight="300" d:DesignWidth="300">
+    <Border Style="{StaticResource CardStyle}" 
+            Margin="0,0,0,15" Padding="16,12" MinHeight="60">
+        <Grid>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="Auto"/>
+                <ColumnDefinition Width="*" MinWidth="200"/>
+                <ColumnDefinition Width="Auto"/>
+            </Grid.ColumnDefinitions>
+            <TextBlock Grid.Column="0" Text="Search:" VerticalAlignment="Center" Margin="0,0,10,0" FontSize="14"/>
+            <TextBox Grid.Column="1" Style="{StaticResource SearchBoxStyle}" 
+                     VerticalAlignment="Center"/>
+            <Button Grid.Column="2" Content="Search" Margin="5,0,0,0" />
+        </Grid>
+    </Border>
+</UserControl>
