+++ TourPlanner.Presentation/Presentation/Controls/SearchBarControl.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace TourPlanner.Presentation.Presentation.Controls;
+
+public partial class SearchBarControl : UserControl
+{
+    public SearchBarControl()
+    {
+        InitializeComponent();
+    }
+}
