+++ TourPlanner.Presentation/Presentation/Controls/TourDetailsControl.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace TourPlanner.Presentation.Presentation.Controls;
+
+public partial class TourDetailsControl : UserControl
+{
+    public TourDetailsControl()
+    {
+        InitializeComponent();
+    }
+}
