+++ TourPlanner.Presentation/Presentation/Controls/TourListControl.xaml
@@ -1,0 +1,81 @@
+﻿<UserControl x:Class="TourPlanner.Presentation.Presentation.Controls.TourListControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+             
+    <Border Style="{StaticResource CardStyle}" Margin="0,0,10,0" Padding="0" MinHeight="400">
+        <DockPanel>
+            <Border DockPanel.Dock="Top" Background="#F5F5F5" BorderBrush="#DDDDDD" 
+                    BorderThickness="0,0,0,1" Padding="12,8" MinHeight="44">
+                <Grid>
+                    <TextBlock Text="Tours" Style="{StaticResource HeaderTextStyle}" 
+                              VerticalAlignment="Center"/>
+                    <StackPanel Orientation="Horizontal" HorizontalAlignment="Right">
+                        <Button Content="+" Style="{StaticResource ActionButtonStyle}" 
+                                ToolTip="Add New Tour" Command="{Binding ShowAddTourCommand}"/>
+                        <Button Content="-" Style="{StaticResource ActionButtonStyle}" 
+                                Background="#F44336" 
+                                ToolTip="Remove Selected Tour" Command="{Binding ShowRemoveTourCommand}"/>
+                        <Button Content="⋮" Style="{StaticResource ActionButtonStyle}" 
+                                ToolTip="More Options"/>
+                    </StackPanel>
+                </Grid>
+            </Border>
+            
+            <ListView Style="{StaticResource ListViewStyle}" Margin="0" BorderThickness="0"
+                      ScrollViewer.HorizontalScrollBarVisibility="Disabled">
+                <ListView.ItemTemplate>
+                    <DataTemplate>
+                        <Border Padding="8" BorderBrush="#EEEEEE" BorderThickness="0,0,0,1" MinWidth="200">
+                            <Grid>
+                                <Grid.RowDefinitions>
+                                    <RowDefinition Height="Auto"/>
+                                    <RowDefinition Height="Auto"/>
+                                </Grid.RowDefinitions>
+                                <Grid.ColumnDefinitions>
+                                    <ColumnDefinition Width="*" MinWidth="150"/>
+                                    <ColumnDefinition Width="Auto"/>
+                                </Grid.ColumnDefinitions>
+                                
+                                <TextBlock Grid.Row="0" Grid.Column="0" Text="Tour Name Example" 
+                                           FontWeight="SemiBold" FontSize="14" 
+                                           TextWrapping="Wrap" TextTrimming="CharacterEllipsis"/>
+                                
+                                <StackPanel Grid.Row="1" Grid.Column="0" Orientation="Horizontal" Margin="0,4,0,0">
+                                    <TextBlock Text="Vienna" Foreground="#666666" Margin="0,0,5,0" TextWrapping="Wrap"/>
+                                    <TextBlock Text="→" Foreground="#666666" Margin="0,0,5,0"/>
+                                    <TextBlock Text="Salzburg" Foreground="#666666" TextWrapping="Wrap"/>
+                                </StackPanel>
+                                
+                                <StackPanel Grid.Row="0" Grid.Column="1" Grid.RowSpan="2" Orientation="Vertical" 
+                                           HorizontalAlignment="Right" VerticalAlignment="Center" MinWidth="40">
+                                    <StackPanel Orientation="Horizontal">
+                                        <TextBlock Text="P:" ToolTip="Popularity" Margin="0,0,2,0" 
+                                                  Foreground="#888888" FontSize="12"/>
+                                        <TextBlock Text="4.2" Foreground="#888888" 
+                                                  FontSize="12"/>
+                                    </StackPanel>
+                                    <StackPanel Orientation="Horizontal">
+                                        <TextBlock Text="CF:" ToolTip="Child-Friendliness" Margin="0,0,2,0" 
+                                                  Foreground="#888888" FontSize="12"/>
+                                        <TextBlock Text="3.5" Foreground="#888888" 
+                                                  FontSize="12"/>
+                                    </StackPanel>
+                                </StackPanel>
+                            </Grid>
+                        </Border>
+                    </DataTemplate>
+                </ListView.ItemTemplate>
+                <!-- Example items for design-time preview -->
+                <ListViewItem>
+                    <TextBlock Text="Vienna to Salzburg" />
+                </ListViewItem>
+                <ListViewItem>
+                    <TextBlock Text="Innsbruck Mountain Trail" />
+                </ListViewItem>
+                <ListViewItem>
+                    <TextBlock Text="Danube Bike Tour" />
+                </ListViewItem>
+            </ListView>
+        </DockPanel>
+    </Border>
+</UserControl>
