+++ TourPlanner.Presentation/Presentation/Controls/TourListControl.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace TourPlanner.Presentation.Presentation.Controls;
+
+public partial class TourListControl : UserControl
+{
+    public TourListControl()
+    {
+        InitializeComponent();
+    }
+}
