+++ TourPlanner.Presentation/Presentation/Controls/TourListControls/TourDescriptionControl.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace TourPlanner.Presentation.Presentation.Controls.TourListControls;
+
+public partial class TourDescriptionControl : UserControl
+{
+    public TourDescriptionControl()
+    {
+        InitializeComponent();
+    }
+}
