+++ TourPlanner.Presentation/Presentation/Controls/TourListControls/TourInformationControl.xaml
@@ -1,0 +1,41 @@
+﻿<UserControl x:Class="TourPlanner.Presentation.Presentation.Controls.TourListControls.TourInformationControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             mc:Ignorable="d"
+             d:DesignHeight="300" d:DesignWidth="300">
+    <Border Style="{StaticResource CardStyle}" Margin="0,0,5,0" Background="#F9F9F9" MinHeight="160">
+        <StackPanel>
+            <TextBlock Text="Tour Information" Style="{StaticResource SubHeaderTextStyle}"/>
+            <Grid Margin="0,5,0,0">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="Auto" MinWidth="100"/>
+                    <ColumnDefinition Width="*" MinWidth="80"/>
+                </Grid.ColumnDefinitions>
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="Auto"/>
+                </Grid.RowDefinitions>
+                
+                <TextBlock Grid.Row="0" Grid.Column="0" Text="From:" FontWeight="SemiBold" Margin="0,0,10,5" TextWrapping="Wrap"/>
+                <TextBlock Grid.Row="0" Grid.Column="1" Text="Vienna" Margin="0,0,0,5" TextWrapping="Wrap"/>
+                
+                <TextBlock Grid.Row="1" Grid.Column="0" Text="To:" FontWeight="SemiBold" Margin="0,0,10,5" TextWrapping="Wrap"/>
+                <TextBlock Grid.Row="1" Grid.Column="1" Text="Salzburg" Margin="0,0,0,5" TextWrapping="Wrap"/>
+                
+                <TextBlock Grid.Row="2" Grid.Column="0" Text="Distance:" FontWeight="SemiBold" Margin="0,0,10,5" TextWrapping="Wrap"/>
+                <TextBlock Grid.Row="2" Grid.Column="1" Text="295 km" Margin="0,0,0,5" TextWrapping="Wrap"/>
+                
+                <TextBlock Grid.Row="3" Grid.Column="0" Text="Transport:" FontWeight="SemiBold" Margin="0,0,10,5" TextWrapping="Wrap"/>
+                <TextBlock Grid.Row="3" Grid.Column="1" Text="Bicycle" Margin="0,0,0,5" TextWrapping="Wrap"/>
+                
+                <TextBlock Grid.Row="4" Grid.Column="0" Text="Est. Time:" FontWeight="SemiBold" Margin="0,0,10,5" TextWrapping="Wrap"/>
+                <TextBlock Grid.Row="4" Grid.Column="1" Text="15 hours" Margin="0,0,0,5" TextWrapping="Wrap"/>
+            </Grid>
+        </StackPanel>
+    </Border>
+</UserControl>
