+++ TourPlanner.Presentation/Presentation/Controls/TourListControls/TourInformationControl.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace TourPlanner.Presentation.Presentation.Controls.TourListControls;
+
+public partial class TourInformationControl : UserControl
+{
+    public TourInformationControl()
+    {
+        InitializeComponent();
+    }
+}
