+++ TourPlanner.Presentation/Presentation/Controls/TourListControls/TourStatisticControl.xaml
@@ -1,0 +1,37 @@
+﻿<UserControl x:Class="TourPlanner.Presentation.Presentation.Controls.TourListControls.TourStatisticControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             mc:Ignorable="d"
+             d:DesignHeight="300" d:DesignWidth="300">
+    <Border Style="{StaticResource CardStyle}" Margin="5,0,0,0" Background="#F9F9F9" MinHeight="160">
+        <StackPanel>
+            <TextBlock Text="Statistics" Style="{StaticResource SubHeaderTextStyle}"/>
+            <Grid Margin="0,5,0,0">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="Auto" MinWidth="130"/>
+                    <ColumnDefinition Width="*" MinWidth="50"/>
+                </Grid.ColumnDefinitions>
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="Auto"/>
+                </Grid.RowDefinitions>
+                
+                <TextBlock Grid.Row="0" Grid.Column="0" Text="Popularity:" FontWeight="SemiBold" Margin="0,0,10,5" TextWrapping="Wrap"/>
+                <TextBlock Grid.Row="0" Grid.Column="1" Text="4.2/5" Margin="0,0,0,5" TextWrapping="Wrap"/>
+                
+                <TextBlock Grid.Row="1" Grid.Column="0" Text="Child-Friendliness:" FontWeight="SemiBold" Margin="0,0,10,5" TextWrapping="Wrap"/>
+                <TextBlock Grid.Row="1" Grid.Column="1" Text="3.5/5" Margin="0,0,0,5" TextWrapping="Wrap"/>
+                
+                <TextBlock Grid.Row="2" Grid.Column="0" Text="Average Rating:" FontWeight="SemiBold" Margin="0,0,10,5" TextWrapping="Wrap"/>
+                <TextBlock Grid.Row="2" Grid.Column="1" Text="4.7/5" Margin="0,0,0,5" TextWrapping="Wrap"/>
+                
+                <TextBlock Grid.Row="3" Grid.Column="0" Text="Total Logs:" FontWeight="SemiBold" Margin="0,0,10,5" TextWrapping="Wrap"/>
+                <TextBlock Grid.Row="3" Grid.Column="1" Text="8" Margin="0,0,0,5" TextWrapping="Wrap"/>
+            </Grid>
+        </StackPanel>
+    </Border>
+</UserControl>
