+++ TourPlanner.Presentation/Presentation/Controls/TourListControls/TourStatisticControl.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace TourPlanner.Presentation.Presentation.Controls.TourListControls;
+
+public partial class TourStatisticControl : UserControl
+{
+    public TourStatisticControl()
+    {
+        InitializeComponent();
+    }
+}
