+++ TourPlanner.Presentation/Presentation/Views/AddingTourWindow.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows;
+
+namespace TourPlanner.Presentation.Presentation.Views;
+
+public partial class AddingTourWindow : Window
+{
+    public AddingTourWindow()
+    {
+        InitializeComponent();
+    }
+}
