+++ TourPlanner.Presentation/Presentation/Views/RemovingTourWindow.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows;
+
+namespace TourPlanner.Presentation.Presentation.Views;
+
+public partial class RemovingTourWindow : Window
+{
+    public RemovingTourWindow()
+    {
+        InitializeComponent();
+    }
+}
